requirements.txt
PyQt6>=6.6.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyqtgraph>=0.13.0
